Plot in R
Mean trend lines and Subject-Specific fits for all 3 models in my final project. ggplot2 is used to create the graphics and most numbers are hard-coded in the script. 

Future edits should import estimates from a file so that the code is more dynamic. For simplicity, I am leaving as is for now.